refactor(api): PR5 — split signals into api/signals.py + db/signals.py
Move signal queries (save_scan, get_scans, get_latest_signal,
get_latest_scan, get_signals_summary) to db/signals.py.

Move 5 signal routes (/signals, /signals/performance, /signals/latest,
/signals/latest/message, /signals/{scan_id}), notification filters
(should_notify_signal, _is_duplicate_signal, _mark_notified — held back
from PR3 because they're signal-domain not telegram-delivery), and CSV/log
appenders (update_symbols_json, append_signal_csv, append_signal_log)
to api/signals.py.

The writing version of check_pending_signal_outcomes stays in btc_api.py
until PR7 — it's coupled to scanner_loop's writing path.

btc_api.py re-exports for scanner_loop and other legacy callers until PR7,
including: route handler refs (list_signals, latest_signal, latest_message,
signal_by_id, get_signals_performance), SIGNALS_LOG_FILE, and the mutable
_notified_signals dict (same object, tested via btc_api._notified_signals).

Heaviest PR of the refactor (~600 LOC moved). Parity test backed by
9-key snapshot baseline covers list filters, latest, performance, by-id.

1010 tests pass.
